clean out experiment factories
